requirements.txt

# Manual insallation:

Flask==3.1.0 # Web framework 
psycopg2-binary==2.9.10 # PostgreSQL driver 
python-dotenv==1.0.1 # Env variable management 


# Flask dependencies, automatically installed
blinker==1.9.0
click==8.1.8
colorama==0.4.6
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
Werkzeug==3.1.3
